mumps: install mpiseq headers for '~mpi'
Prior to this commit, spack installs a library called `libmpiseq` into
`spec['mumps'].prefix.lib` when it builds MUMPS without MPI. However, it
does not also install the headers corresponding to this library, so it
is impossible to compile source files that depend on function calls
implemented in this library. This commit fixes this problem by
installing these headers, which are needed by packages (e.g., IPOPT)
that depend on these headers.