Add configparser v3.5.1, fixes backports namespace (#10433)
Fixes #8343 #4154 #7370
Closes #4155 

All but one of the backports packages were using "pkgutil-style"
namespacing, but configparser was using "pkg_resource-style".
A namespace needs to be one style or another. This adds a new
configparser version which switches to using the pkgutil-style
namespace approach.

With this update, one can now install flake8 with Spack and use
it by loading modules.