ci: fix inconsistent behavior for packages with maintainers not in org (#1527)
The GitHub API appears to be nondeterministic in how it responds to reviewer requests when a subset of the reviewers don't have triage permissions for a repository. Sometimes it'll add all of the users with permissions and respond 422 and sometimes it'll respond just with the error status code without adding anyone.

This PR modifies the existing approach so that we first verify the permissions for all maintainers before we attempt to request their review