chapel: add 2.6 (#1483)
* Add placeholder version entry for Chapel 2.6

* Restore ROCm support Chapel version requirement

Undoing apparent unintended semantic change in
2bf4ab9585c8d483cc8581d65912703d3f020393

* Add tmp tarball URL and SHA256

* Require Chapel 2.2 or later for ROCm support

* Require bundled LLVM for ROCm support

* Rephrase llvm=bundled requirement when using ROCm

* Rephrase ROCm requirement again and add comment

* Wrap long comment lines

* Remove tmp tarball URL in favor of default

* Deprecate versions @2.0:2.3

* Force use of an allowed compiler when +rocm

* Derive non-rocmcc compiler list from compiler_map

* Add dropped % at beginning of compiler names

* Adjust ROCm compiler requirement policy

`%clang` and `%llvm` are aliases as far as Spack is concerned, so
depending on one means depending on both. Both appear in the `compiler_map`,
so the policy `any_of` allows us to simultaneously depend on both.

* print CHPL_ envvars during the build

* Fix chapel %clang

Spack %clang is a synonym for %llvm, and both should result in CHPL_HOST_COMPILER=clang, not llvm

* Change single quotes to double quotes

* Add missing entries in chpl_env_vars

These missing entries (in particular `CHPL_MAKE_THIRD_PARTY`) was
allowing the package run environment to contaminate a subsequent
attempt to install the package (in this case, leading to a missing
Makefile error in the install step).

---------

Co-authored-by: bonachea <dobonachea@lbl.gov>